wip : new project setup to include testing
