+++ main.cpp
@@ -1,6 +1,6 @@
-#include "utils/logger.hpp"
+#include "logger.hpp"
 
-int main(){
-    Logger::INFO("Motors", "Speed: {:.2f} RPM", 123.456);
+int main() {
+    newton::Logger::INFO("Motors", "Speed: {:.2f} RPM", 123.456);
     return 0;
 }
