+++ utils/tests/test_logger.cpp
@@ -1,0 +1,33 @@
+#include <catch2/catch_test_macros.hpp>
+#include "logger.hpp"
+#include <fstream>
+#include <string>
+
+using namespace newton;
+
+TEST_CASE("Logger basic functionality", "[utils]") {
+  auto& logger = Logger::get_instance();
+  logger.set_logfile("test.log");
+
+  SECTION("info logging") {
+    logger.enable(Logger::Level::INFO);
+    newton::Logger::INFO("Test", "Info message");
+
+
+    std::ifstream file("test.log");
+    std::string line;
+    std::getline(file, line);
+    REQUIRE(line.find("INFO") != std::string::npos);
+    REQUIRE(line.find("Info message") != std::string::npos);
+  }
+
+  SECTION("debug messages are filtered when info level") {
+    logger.enable(Logger::Level::INFO);
+    Logger::DEBUG("Test", "Debug message");
+
+    std::ifstream file("test.log");
+    std::string line;
+    std::getline(file, line);
+    REQUIRE(line.find("Debug message") == std::string::npos);
+  }
+}
